+++ src/gfx/CMakeLists.txt
@@ -17,10 +17,10 @@ add_library(gfx
   utils.cpp
   feature.cpp
   texture.cpp
-  helpers/wireframe.cpp
-  helpers/shapes.cpp
-  helpers/gizmo_input.cpp
-  helpers/gizmos.cpp
+  gizmos/wireframe.cpp
+  gizmos/shapes.cpp
+  gizmos/gizmo_input.cpp
+  gizmos/gizmos.cpp
   shader/generator.cpp
   shader/blocks.cpp
   shader/types.cpp
