+++ src/gfx/gizmos/translation_gizmo.hpp
